Add white noise example
This one is interesting in that it clearly shows the channel
ordering assumptions at least on my NVIDIA GPU. Specifically,
the image dimensions are specified as HWCN, but the original
code had WHCN instead. Need to check whether this can be inferred
at runtime or is just an artifact of an API change ... that
would be weird though.