Fix a bug in conway.rs
Due to the addition of bool values in the state propagation,
cell values could be higher than one. As a result, such a cell would
be counted multiple times as a neighbor due to the convolution.
The issue was fixed by adding a clamping step ensuring that values
would always be in range 0..1.